(ref) rename messages file to receive-sqs
