+++ service/src/server/helpers/receive-sqs.js
